make app usable even if geolocation stalls
by setting the map's location right away and then setting it
again later if geolocation works.  Not perfect (what if the user
already started looking around; maybe we should detect that and
if they do, refrain from moving the map location by geolocation.
Maybe.).  Probably an improvement though!

(deals with an issue that sometimes happens in some of my browsers)